refactor: split decoder.go and reflect_encode.go by responsibility
Pure file-level split within the root package — no symbol, API, or wire
change. The two largest files were doing too much in one place.

  decoder.go (1248 LOC) -> decoder.go (core: type, ctors, setup, header)
                         + decoder_read.go (typed Read* primitives)
                         + decoder_skip.go (Skip)
  reflect_encode.go (1296 LOC) -> reflect_encode.go (per-kind enc/dec)
                                + reflect_desc.go (type-descriptor build)

Verified with build + race + golden (byte-identical) under both build
tags, golangci-lint clean, and arm64 golden + decode under emulation.
