+++ decoder.go
@@ -1,11 +1,8 @@
 package qdf
 
 import (
-	"math"
-
 	"github.com/alex60217101990/qdf/internal/intern"
 	"github.com/alex60217101990/qdf/internal/rans"
-	"github.com/alex60217101990/qdf/internal/unsafestr"
 )
 
 // Decoder reads QDF wire data from a single input buffer. Call SetInput
@@ -185,1064 +182,4 @@ func (d *Decoder) next() (byte, error) {
 	}
 	d.i++
 	return t, nil
-}
-
-// ----- primitives -----
-
-func (d *Decoder) ReadNil() error {
-	t, err := d.next()
-	if err != nil {
-		return err
-	}
-	if t != tagNil {
-		return ErrTypeMismatch
-	}
-	return nil
-}
-
-func (d *Decoder) ReadBool() (bool, error) {
-	t, err := d.next()
-	if err != nil {
-		return false, err
-	}
-	switch t {
-	case tagTrue:
-		return true, nil
-	case tagFalse:
-		return false, nil
-	default:
-		return false, ErrTypeMismatch
-	}
-}
-
-// ReadUint reads a value that was encoded with WriteUint or WriteInt with a
-// non-negative argument. Returns an error if the next value is negative.
-func (d *Decoder) ReadUint() (uint64, error) {
-	t, err := d.next()
-	if err != nil {
-		return 0, err
-	}
-	return d.decodeUint(t)
-}
-
-func (d *Decoder) decodeUint(t byte) (uint64, error) {
-	if t <= tagFixintMax {
-		return uint64(t), nil
-	}
-	switch t {
-	case tagUint8:
-		if d.i >= len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		v := uint64(d.buf[d.i])
-		d.i++
-		return v, nil
-	case tagUint16:
-		if d.i+2 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		v := uint64(readU16(d.buf[d.i:]))
-		d.i += 2
-		return v, nil
-	case tagUint32:
-		if d.i+4 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		v := uint64(readU32(d.buf[d.i:]))
-		d.i += 4
-		return v, nil
-	case tagUint64:
-		if d.i+8 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		v := readU64(d.buf[d.i:])
-		d.i += 8
-		return v, nil
-	}
-	return 0, ErrTypeMismatch
-}
-
-func (d *Decoder) ReadInt() (int64, error) {
-	t, err := d.next()
-	if err != nil {
-		return 0, err
-	}
-	return d.decodeInt(t)
-}
-
-func (d *Decoder) decodeInt(t byte) (int64, error) {
-	// fixint positive
-	if t <= tagFixintMax {
-		return int64(t), nil
-	}
-	// negfixint range 0xD8..0xDF
-	if t >= tagNegfixint && t <= tagNegfixint|tagNegfixintMask {
-		return -int64(t&tagNegfixintMask) - 1, nil
-	}
-	switch t {
-	case tagInt8:
-		if d.i >= len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		v := int64(int8(d.buf[d.i]))
-		d.i++
-		return v, nil
-	case tagInt16:
-		if d.i+2 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		v := int64(int16(readU16(d.buf[d.i:])))
-		d.i += 2
-		return v, nil
-	case tagInt32:
-		if d.i+4 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		v := int64(int32(readU32(d.buf[d.i:])))
-		d.i += 4
-		return v, nil
-	case tagInt64:
-		if d.i+8 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		v := int64(readU64(d.buf[d.i:]))
-		d.i += 8
-		return v, nil
-	}
-	// fall back to the unsigned decoder for tagUintN
-	u, err := d.decodeUint(t)
-	if err != nil {
-		return 0, err
-	}
-	if u > math.MaxInt64 {
-		return 0, ErrTypeMismatch
-	}
-	return int64(u), nil
-}
-
-func (d *Decoder) ReadFloat32() (float32, error) {
-	t, err := d.next()
-	if err != nil {
-		return 0, err
-	}
-	if t != tagFloat32 {
-		return 0, ErrTypeMismatch
-	}
-	if d.i+4 > len(d.buf) {
-		return 0, ErrShortBuffer
-	}
-	v := math.Float32frombits(readU32(d.buf[d.i:]))
-	d.i += 4
-	return v, nil
-}
-
-func (d *Decoder) ReadFloat64() (float64, error) {
-	t, err := d.next()
-	if err != nil {
-		return 0, err
-	}
-	switch t {
-	case tagFloat64:
-		if d.i+8 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		v := math.Float64frombits(readU64(d.buf[d.i:]))
-		d.i += 8
-		return v, nil
-	case tagFloat32:
-		if d.i+4 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		v := float64(math.Float32frombits(readU32(d.buf[d.i:])))
-		d.i += 4
-		return v, nil
-	}
-	return 0, ErrTypeMismatch
-}
-
-// ReadString returns the next string. If the decoder is in noCopy mode the
-// returned string aliases the input buffer; otherwise a copy is made.
-//
-// When the tag at the cursor resolves through the intern table —
-// tagInternStr, tagStateRef, tagStateMTF, tagStatePair, tagStateRepeat —
-// readStringBytes leaves d.state.lastID set to the entry it just
-// touched. We return the pre-materialised d.state.stringValues[id]
-// from that path instead of allocating a fresh `string(b)` copy on
-// every state-ref hit. Inline reads (fixstr, str8/16/32) set
-// lastID = lruInvalidID, fall through, and pay the copy as before.
-func (d *Decoder) ReadString() (string, error) {
-	b, err := d.readStringBytes()
-	if err != nil {
-		return "", err
-	}
-	if d.noCopy {
-		return unsafestr.String(b), nil
-	}
-	if d.state != nil && d.state.lastID != lruInvalidID {
-		if s, ok := d.state.getString(d.state.lastID); ok {
-			return s, nil
-		}
-	}
-	return string(b), nil
-}
-
-// readStringBytes returns the raw bytes of a string/bin value without
-// allocating; the returned slice aliases the input buffer or the decoder
-// state table.
-func (d *Decoder) readStringBytes() ([]byte, error) {
-	t, err := d.next()
-	if err != nil {
-		return nil, err
-	}
-	// Inline (non-intern) string / bin reads break the Markov-0 chain so
-	// a subsequent tagStateRepeat cannot resurrect a stale state-ref
-	// from before the inline emission. State-ref and intern branches
-	// below restore the invariant explicitly.
-	invalidateLast := d.state != nil
-	// fixstr
-	if t >= tagFixstr && t <= tagFixstr|tagFixstrMask {
-		n := int(t & tagFixstrMask)
-		if d.i+n > len(d.buf) {
-			return nil, ErrShortBuffer
-		}
-		out := d.buf[d.i : d.i+n]
-		d.i += n
-		if invalidateLast {
-			d.state.lastID = lruInvalidID
-		}
-		return out, nil
-	}
-	switch t {
-	case tagStr8, tagBin8:
-		if d.i >= len(d.buf) {
-			return nil, ErrShortBuffer
-		}
-		n := int(d.buf[d.i])
-		d.i++
-		if d.i+n > len(d.buf) {
-			return nil, ErrShortBuffer
-		}
-		out := d.buf[d.i : d.i+n]
-		d.i += n
-		if invalidateLast {
-			d.state.lastID = lruInvalidID
-		}
-		return out, nil
-	case tagStr16, tagBin16:
-		if d.i+2 > len(d.buf) {
-			return nil, ErrShortBuffer
-		}
-		n := int(readU16(d.buf[d.i:]))
-		d.i += 2
-		if d.i+n > len(d.buf) {
-			return nil, ErrShortBuffer
-		}
-		out := d.buf[d.i : d.i+n]
-		d.i += n
-		if invalidateLast {
-			d.state.lastID = lruInvalidID
-		}
-		return out, nil
-	case tagStr32, tagBin32:
-		if d.i+4 > len(d.buf) {
-			return nil, ErrShortBuffer
-		}
-		n := int(readU32(d.buf[d.i:]))
-		d.i += 4
-		if d.i+n > len(d.buf) {
-			return nil, ErrShortBuffer
-		}
-		out := d.buf[d.i : d.i+n]
-		d.i += n
-		if invalidateLast {
-			d.state.lastID = lruInvalidID
-		}
-		return out, nil
-	case tagInternStr, tagInternBin:
-		// Read length-prefixed payload, then register it in the state table.
-		n64, n := readUvarint(d.buf[d.i:])
-		if n <= 0 {
-			return nil, ErrInvalidLength
-		}
-		d.i += n
-		// Length validated in uint64 to avoid the int sign-bit pitfall on a
-		// hostile varint. A 10-byte varint can encode values up to 2^64-1;
-		// our buffer is always < 2^63, so the uint64 comparison is safe.
-		if n64 > uint64(len(d.buf)-d.i) {
-			return nil, ErrShortBuffer
-		}
-		nn := int(n64)
-		out := d.buf[d.i : d.i+nn]
-		d.i += nn
-		if d.state == nil {
-			d.state = newDecState()
-		}
-		// Register the bytes as-is; they alias the input. If the caller
-		// later turns this into a copy, the table still references the alias
-		// which is fine for the lifetime of the buffer.
-		id := d.state.append(out)
-		if d.state.lastID != lruInvalidID {
-			d.state.pairRecord(d.state.lastID, id)
-		}
-		d.state.lastID = id
-		return out, nil
-	case tagStateRef:
-		id64, n := readUvarint(d.buf[d.i:])
-		if n <= 0 {
-			return nil, ErrInvalidLength
-		}
-		d.i += n
-		if d.state == nil {
-			return nil, ErrUnknownStateID
-		}
-		out, ok := d.state.get(uint32(id64))
-		if !ok {
-			return nil, ErrUnknownStateID
-		}
-		// Mirror encoder's LRU update so a subsequent tagStateMTF
-		// resolves to the same ID position.
-		d.state.lruMoveToFront(uint32(id64))
-		if d.state.lastID != lruInvalidID {
-			d.state.pairRecord(d.state.lastID, uint32(id64))
-		}
-		d.state.lastID = uint32(id64)
-		return out, nil
-	case tagStateMTF:
-		rank64, n := readUvarint(d.buf[d.i:])
-		if n <= 0 {
-			return nil, ErrInvalidLength
-		}
-		d.i += n
-		if d.state == nil {
-			return nil, ErrUnknownStateID
-		}
-		// Side-cache lookup: the encoder emits tagStateMTF only when
-		// rank fits in fewer varuint bytes than the raw id, which
-		// caps rank at mruRingSize-1 for the common 2-byte id case.
-		// The MRU ring resolves that range in O(1) without walking
-		// the LRU chain. Older/larger ranks fall back to the chain
-		// walk for forward-compatibility.
-		rank := uint32(rank64)
-		id, ok := d.state.mruIDAtRank(rank)
-		if !ok {
-			id, ok = d.state.lruIDAtRank(rank)
-		}
-		if !ok {
-			return nil, ErrUnknownStateID
-		}
-		out, ok := d.state.get(id)
-		if !ok {
-			return nil, ErrUnknownStateID
-		}
-		d.state.lruMoveToFront(id)
-		if d.state.lastID != lruInvalidID {
-			d.state.pairRecord(d.state.lastID, id)
-		}
-		d.state.lastID = id
-		return out, nil
-	case tagStatePair:
-		if d.state == nil || d.state.lastID == lruInvalidID {
-			return nil, ErrUnknownStateID
-		}
-		rank64, n := readUvarint(d.buf[d.i:])
-		if n <= 0 {
-			return nil, ErrInvalidLength
-		}
-		d.i += n
-		if rank64 >= pairPredK {
-			return nil, ErrUnknownStateID
-		}
-		prev := d.state.lastID
-		id, ok := d.state.pairAtRank(prev, uint8(rank64))
-		if !ok {
-			return nil, ErrUnknownStateID
-		}
-		out, ok := d.state.get(id)
-		if !ok {
-			return nil, ErrUnknownStateID
-		}
-		// Same post-emit bookkeeping as tagStateRef so the encoder
-		// and decoder mirror chains diverge nowhere.
-		d.state.lruMoveToFront(id)
-		d.state.pairRecord(prev, id)
-		d.state.lastID = id
-		return out, nil
-	case tagStateRepeat:
-		if d.state == nil || d.state.lastID == lruInvalidID {
-			return nil, ErrUnknownStateID
-		}
-		out, ok := d.state.get(d.state.lastID)
-		if !ok {
-			return nil, ErrUnknownStateID
-		}
-		// Pair predictor: mirror encoder's self-record.
-		d.state.pairRecord(d.state.lastID, d.state.lastID)
-		return out, nil
-	}
-	return nil, ErrTypeMismatch
-}
-
-// ReadBytes returns a []byte value. With noCopy = true the result aliases
-// the input.
-func (d *Decoder) ReadBytes() ([]byte, error) {
-	b, err := d.readStringBytes()
-	if err != nil {
-		return nil, err
-	}
-	if d.noCopy {
-		return b, nil
-	}
-	out := make([]byte, len(b))
-	copy(out, b)
-	return out, nil
-}
-
-// ReadArrayHeader returns the element count.
-func (d *Decoder) ReadArrayHeader() (int, error) {
-	t, err := d.next()
-	if err != nil {
-		return 0, err
-	}
-	if t >= tagFixarr && t <= tagFixarr|tagFixarrMask {
-		return int(t & tagFixarrMask), nil
-	}
-	switch t {
-	case tagArr16:
-		if d.i+2 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		n := int(readU16(d.buf[d.i:]))
-		d.i += 2
-		return n, nil
-	case tagArr32:
-		if d.i+4 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		n := int(readU32(d.buf[d.i:]))
-		d.i += 4
-		return n, nil
-	}
-	return 0, ErrTypeMismatch
-}
-
-// ReadMapHeader returns the key/value pair count.
-func (d *Decoder) ReadMapHeader() (int, error) {
-	t, err := d.next()
-	if err != nil {
-		return 0, err
-	}
-	switch t {
-	case tagMap8:
-		if d.i >= len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		n := int(d.buf[d.i])
-		d.i++
-		return n, nil
-	case tagMap16:
-		if d.i+2 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		n := int(readU16(d.buf[d.i:]))
-		d.i += 2
-		return n, nil
-	case tagMap32:
-		if d.i+4 > len(d.buf) {
-			return 0, ErrShortBuffer
-		}
-		n := int(readU32(d.buf[d.i:]))
-		d.i += 4
-		return n, nil
-	}
-	return 0, ErrTypeMismatch
-}
-
-func (d *Decoder) ReadTimestampNano() (int64, error) {
-	t, err := d.next()
-	if err != nil {
-		return 0, err
-	}
-	if t != tagTimestamp {
-		return 0, ErrTypeMismatch
-	}
-	if d.i+8 > len(d.buf) {
-		return 0, ErrShortBuffer
-	}
-	v := int64(readU64(d.buf[d.i:]))
-	d.i += 8
-	return v, nil
-}
-
-// Skip advances past one value without materializing it.
-func (d *Decoder) Skip() error {
-	t, err := d.peekTag()
-	if err != nil {
-		return err
-	}
-	switch {
-	case t <= tagFixintMax:
-		d.i++
-		return nil
-	case t >= tagFixstr && t <= tagFixstr|tagFixstrMask:
-		n := int(t & tagFixstrMask)
-		if d.i+1+n > len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i += 1 + n
-		return nil
-	case t >= tagFixarr && t <= tagFixarr|tagFixarrMask:
-		n := int(t & tagFixarrMask)
-		d.i++
-		for range n {
-			if err := d.Skip(); err != nil {
-				return err
-			}
-		}
-		return nil
-	case t >= tagNegfixint && t <= tagNegfixint|tagNegfixintMask:
-		d.i++
-		return nil
-	}
-	switch t {
-	case tagNil, tagTrue, tagFalse:
-		d.i++
-		return nil
-	case tagUint8, tagInt8:
-		if d.i+2 > len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i += 2
-		return nil
-	case tagUint16, tagInt16:
-		if d.i+3 > len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i += 3
-		return nil
-	case tagUint32, tagInt32, tagFloat32:
-		if d.i+5 > len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i += 5
-		return nil
-	case tagUint64, tagInt64, tagFloat64, tagTimestamp:
-		if d.i+9 > len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i += 9
-		return nil
-	case tagStr8, tagBin8, tagMap8:
-		// Map8 is a count, not a byte length; handle separately below.
-		if t == tagMap8 {
-			n, err := d.ReadMapHeader()
-			if err != nil {
-				return err
-			}
-			for range n {
-				if err := d.Skip(); err != nil {
-					return err
-				}
-				if err := d.Skip(); err != nil {
-					return err
-				}
-			}
-			return nil
-		}
-		d.i++ // tag
-		if d.i >= len(d.buf) {
-			return ErrShortBuffer
-		}
-		n := int(d.buf[d.i])
-		d.i++
-		if d.i+n > len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i += n
-		return nil
-	case tagStr16, tagBin16, tagArr16, tagMap16:
-		d.i++
-		if t == tagArr16 {
-			if d.i+2 > len(d.buf) {
-				return ErrShortBuffer
-			}
-			n := int(readU16(d.buf[d.i:]))
-			d.i += 2
-			for range n {
-				if err := d.Skip(); err != nil {
-					return err
-				}
-			}
-			return nil
-		}
-		if t == tagMap16 {
-			if d.i+2 > len(d.buf) {
-				return ErrShortBuffer
-			}
-			n := int(readU16(d.buf[d.i:]))
-			d.i += 2
-			for range n {
-				if err := d.Skip(); err != nil {
-					return err
-				}
-				if err := d.Skip(); err != nil {
-					return err
-				}
-			}
-			return nil
-		}
-		if d.i+2 > len(d.buf) {
-			return ErrShortBuffer
-		}
-		n := int(readU16(d.buf[d.i:]))
-		d.i += 2
-		if d.i+n > len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i += n
-		return nil
-	case tagStr32, tagBin32, tagArr32, tagMap32:
-		d.i++
-		if t == tagArr32 {
-			if d.i+4 > len(d.buf) {
-				return ErrShortBuffer
-			}
-			n := int(readU32(d.buf[d.i:]))
-			d.i += 4
-			for range n {
-				if err := d.Skip(); err != nil {
-					return err
-				}
-			}
-			return nil
-		}
-		if t == tagMap32 {
-			if d.i+4 > len(d.buf) {
-				return ErrShortBuffer
-			}
-			n := int(readU32(d.buf[d.i:]))
-			d.i += 4
-			for range n {
-				if err := d.Skip(); err != nil {
-					return err
-				}
-				if err := d.Skip(); err != nil {
-					return err
-				}
-			}
-			return nil
-		}
-		if d.i+4 > len(d.buf) {
-			return ErrShortBuffer
-		}
-		n := int(readU32(d.buf[d.i:]))
-		d.i += 4
-		if d.i+n > len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i += n
-		return nil
-	case tagInternStr, tagInternBin:
-		// Read+register; Skip semantics still need the state table to stay
-		// in sync with the stream.
-		_, err := d.readStringBytes()
-		return err
-	case tagStateRef, tagStateRepeat, tagStateMTF, tagStatePair:
-		_, err := d.readStringBytes()
-		return err
-	case tagMapShape:
-		// Wire form mirrors the decode path: either a declaration
-		// (shapeID==0, varuint(N), N keys, N values) or a reuse
-		// (shapeID>0, looked-up N, N values). Skipping must still
-		// advance the shape table on declaration so subsequent
-		// references stay consistent.
-		d.i++
-		shapeID, n := readUvarint(d.buf[d.i:])
-		if n <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += n
-		if d.state == nil {
-			d.state = newDecState()
-		}
-		var cnt int
-		if shapeID == 0 {
-			cnt64, n := readUvarint(d.buf[d.i:])
-			if n <= 0 {
-				return ErrInvalidLength
-			}
-			d.i += n
-			if err := d.CheckLength(int(cnt64), 1); err != nil {
-				return err
-			}
-			cnt = int(cnt64)
-			sh := d.state.shapeDeclare()
-			sh.keyIDs = make([]uint32, 0, cnt)
-			sh.names = make([]string, 0, cnt)
-			for i := 0; i < cnt; i++ {
-				kb, err := d.readStringBytes()
-				if err != nil {
-					return err
-				}
-				sh.names = append(sh.names, string(kb))
-				if d.state.lastID != lruInvalidID {
-					sh.keyIDs = append(sh.keyIDs, d.state.lastID)
-				} else {
-					sh.keyIDs = append(sh.keyIDs, 0)
-				}
-			}
-		} else {
-			sh := d.state.shapeLookup(uint32(shapeID))
-			if sh == nil {
-				return ErrUnknownStateID
-			}
-			cnt = len(sh.names)
-		}
-		for i := 0; i < cnt; i++ {
-			if err := d.Skip(); err != nil {
-				return err
-			}
-		}
-		return nil
-	case tagPackBool:
-		d.i++
-		n64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		// Validate the element count in uint64 BEFORE the signed cast so
-		// a hostile varuint cannot drive nBytes negative and corrupt
-		// d.i. n64 elements need ceil(n64/8) bytes.
-		rem := uint64(len(d.buf) - d.i)
-		if n64 > rem*8 {
-			return ErrShortBuffer
-		}
-		d.i += int((n64 + 7) >> 3)
-		return nil
-	case tagPackRaw:
-		d.i++
-		if d.i >= len(d.buf) {
-			return ErrShortBuffer
-		}
-		k := d.buf[d.i]
-		d.i++
-		w := qpackRawWidthBytes(k)
-		if w == 0 {
-			return ErrBadTag
-		}
-		n64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		if n64 > uint64(len(d.buf)-d.i)/uint64(w) {
-			return ErrShortBuffer
-		}
-		d.i += int(n64) * w
-		return nil
-	case tagPackFor:
-		d.i++
-		if d.i+2 > len(d.buf) {
-			return ErrShortBuffer
-		}
-		// kind, bits
-		d.i++ // skip kind
-		bitsPer := int(d.buf[d.i])
-		d.i++
-		if bitsPer > qpackForMaxBits {
-			return ErrBadTag
-		}
-		// min varuint
-		_, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		n64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		// Same overflow guard as tagPackBool. n64 elements times
-		// bitsPer bits => ceil/8 bytes; validate in uint64 to avoid the
-		// sign-bit pitfall on a hostile varuint.
-		rem := uint64(len(d.buf) - d.i)
-		if bitsPer > 0 && n64 > rem*8/uint64(bitsPer) {
-			return ErrShortBuffer
-		}
-		d.i += int((n64*uint64(bitsPer) + 7) / 8)
-		return nil
-	case tagPackGorilla:
-		d.i++
-		if d.i >= len(d.buf) {
-			return ErrShortBuffer
-		}
-		k := d.buf[d.i]
-		d.i++
-		w := qpackRawWidthBytes(k)
-		if w != 4 && w != 8 {
-			return ErrBadTag
-		}
-		n64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		if n64 == 0 {
-			return nil
-		}
-		if d.i+w > len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i += w
-		// numBits varuint
-		nb64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		rem := uint64(len(d.buf) - d.i)
-		if nb64 > rem*8 {
-			return ErrShortBuffer
-		}
-		d.i += int((nb64 + 7) >> 3)
-		return nil
-	case tagPackALP:
-		d.i++
-		if d.i >= len(d.buf) {
-			return ErrShortBuffer
-		}
-		if d.buf[d.i] != qpackKindFloat64 {
-			return ErrTypeMismatch
-		}
-		d.i++ // kind
-		n64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		if n64 == 0 {
-			return nil
-		}
-		if n64 > alpMaxElems {
-			return ErrInvalidLength
-		}
-		if d.i >= len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i++                                           // d exponent
-		if _, nr := readUvarint(d.buf[d.i:]); nr <= 0 { // forMin
-			return ErrInvalidLength
-		} else {
-			d.i += nr
-		}
-		if d.i >= len(d.buf) {
-			return ErrShortBuffer
-		}
-		width := int(d.buf[d.i])
-		d.i++
-		if width > qpackForMaxBits {
-			return ErrBadTag
-		}
-		if width > 0 {
-			rem := uint64(len(d.buf) - d.i)
-			if n64 > rem*8/uint64(width) {
-				return ErrShortBuffer
-			}
-			d.i += int((n64*uint64(width) + 7) / 8)
-		}
-		excN, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		if excN > n64 {
-			return ErrInvalidLength
-		}
-		for range excN {
-			_, nr := readUvarint(d.buf[d.i:])
-			if nr <= 0 {
-				return ErrInvalidLength
-			}
-			d.i += nr
-			if d.i+8 > len(d.buf) {
-				return ErrShortBuffer
-			}
-			d.i += 8
-		}
-		return nil
-	case tagPackDeltaFor:
-		d.i++
-		if d.i+2 > len(d.buf) {
-			return ErrShortBuffer
-		}
-		d.i++ // skip kind
-		bitsPer := int(d.buf[d.i])
-		d.i++
-		if bitsPer > qpackForMaxBits {
-			return ErrBadTag
-		}
-		// firstVal varuint
-		_, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		// minDelta varuint
-		_, nr = readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		n64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		if n64 >= 2 {
-			// Body holds (n-1) elements at bitsPer each. Compute the
-			// byte size in uint64 to keep the bounds check overflow-safe.
-			bodyBits := (n64 - 1) * uint64(bitsPer)
-			bodyBytes := (bodyBits + 7) >> 3
-			if bodyBytes > uint64(len(d.buf)-d.i) {
-				return ErrShortBuffer
-			}
-			d.i += int(bodyBytes)
-		}
-		return nil
-	case tagPackRLE:
-		d.i++
-		if d.i >= len(d.buf) {
-			return ErrShortBuffer
-		}
-		k := d.buf[d.i]
-		d.i++
-		if k != qpackKindUint64 && k != qpackKindInt64 {
-			return ErrBadTag
-		}
-		n64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		// Body is a sequence of (value-varuint, runLen-varuint)
-		// pairs whose runLen sum equals n. Walk until n elements
-		// consumed; the loop also catches truncated bodies.
-		var produced uint64
-		for produced < n64 {
-			_, nr := readUvarint(d.buf[d.i:])
-			if nr <= 0 {
-				return ErrInvalidLength
-			}
-			d.i += nr
-			runLen, nr := readUvarint(d.buf[d.i:])
-			if nr <= 0 {
-				return ErrInvalidLength
-			}
-			d.i += nr
-			if runLen == 0 || produced+runLen > n64 {
-				return ErrInvalidLength
-			}
-			produced += runLen
-		}
-		return nil
-	case tagPackDict:
-		d.i++
-		if d.i >= len(d.buf) {
-			return ErrShortBuffer
-		}
-		k := d.buf[d.i]
-		d.i++
-		if k != qpackKindUint64 && k != qpackKindInt64 {
-			return ErrBadTag
-		}
-		distinct64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		if distinct64 == 0 || distinct64 > qpackDictMaxDistinct {
-			return ErrBadTag
-		}
-		for range distinct64 {
-			_, nr := readUvarint(d.buf[d.i:])
-			if nr <= 0 {
-				return ErrInvalidLength
-			}
-			d.i += nr
-		}
-		n64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		bp := bitsForDistinct(int(distinct64))
-		if bp == 0 {
-			return nil
-		}
-		rem := uint64(len(d.buf) - d.i)
-		if n64 > rem*8/uint64(bp) {
-			return ErrShortBuffer
-		}
-		d.i += int((n64*uint64(bp) + 7) / 8)
-		return nil
-	case tagPackPFor:
-		d.i++
-		if d.i >= len(d.buf) {
-			return ErrShortBuffer
-		}
-		k := d.buf[d.i]
-		d.i++
-		if k != qpackKindUint64 && k != qpackKindInt64 {
-			return ErrBadTag
-		}
-		n64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		if d.i >= len(d.buf) {
-			return ErrShortBuffer
-		}
-		b := int(d.buf[d.i])
-		d.i++
-		if b > qpackForMaxBits {
-			return ErrBadTag
-		}
-		// min varuint
-		if _, nr := readUvarint(d.buf[d.i:]); nr <= 0 {
-			return ErrInvalidLength
-		} else {
-			d.i += nr
-		}
-		// body: n*b bits
-		rem := uint64(len(d.buf) - d.i)
-		if b > 0 && n64 > rem*8/uint64(b) {
-			return ErrShortBuffer
-		}
-		d.i += int((n64*uint64(b) + 7) / 8)
-		// exception list: excN pairs of (dPos, delta) varuints
-		excN64, nr := readUvarint(d.buf[d.i:])
-		if nr <= 0 {
-			return ErrInvalidLength
-		}
-		d.i += nr
-		if excN64 > n64 {
-			return ErrInvalidLength
-		}
-		for range excN64 {
-			if _, nr := readUvarint(d.buf[d.i:]); nr <= 0 {
-				return ErrInvalidLength
-			} else {
-				d.i += nr
-			}
-			if _, nr := readUvarint(d.buf[d.i:]); nr <= 0 {
-				return ErrInvalidLength
-			} else {
-				d.i += nr
-			}
-		}
-		return nil
-	}
-	return ErrBadTag
 }
